migrations: Ensure all columns with a track id are bigint
Must be run before any of the foreign key patches are run.

+++ configs/db/tables/gems__comm_jobs.300.sql
@@ -50,7 +50,7 @@ CREATE TABLE if not exists gems__comm_jobs (
         gcj_target                  tinyint(1) NOT NULL DEFAULT '0',
         gcj_target_group            varchar(100) CHARACTER SET 'utf8mb4' COLLATE 'utf8mb4_unicode_ci' null,
         gcj_id_organization         bigint unsigned null references gems__organizations (gor_id_organization),
-        gcj_id_track                int unsigned null references gems__tracks (gtr_id_track),
+        gcj_id_track                bigint unsigned null references gems__tracks (gtr_id_track),
         gcj_round_description       varchar(100) CHARACTER SET 'utf8mb4' COLLATE 'utf8mb4_unicode_ci' null,
         gcj_id_survey               int unsigned null references gems__surveys (gsu_id_survey),
