+++ configs/db/tables/gems__track_fields.40.sql
@@ -1,7 +1,7 @@
 
 CREATE TABLE if not exists gems__track_fields (
         gtf_id_field            bigint unsigned not null auto_increment,
-        gtf_id_track            int unsigned not null references gems__tracks (gtr_id_track),
+        gtf_id_track            bigint unsigned not null references gems__tracks (gtr_id_track),
 
         gtf_id_order            int not null default 10,
